fix memory leak

+++ lib/utils/mod.rs
@@ -11,7 +11,7 @@ pub const CHUNK_SIZE: u64 = 1 << 26;
 pub const BLOCK_SIZE: u64 = 1 << 22;
 pub const FS_BLK_SIZE: u64 = 128 << 20;
 pub const FS_FUSE_MAX_IO_SIZE: u64 = 128u64 << 10;
-pub const FS_TOTAL_INODES: u32 = 1 << 20;
+pub const FS_TOTAL_INODES: u64 = 1 << 20;
 pub const FS_META_CACHE_SIZE: usize = 16384;
 
 pub const FS_PAGE_SIZE: u64 = 4096;
